pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "temporalos"
version = "0.1.0"
description = "Video → Structured Decision Intelligence Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "openai>=1.50.0",
    "anthropic>=0.35.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "tenacity>=8.5.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
audio = [
    "faster-whisper>=1.0.3",
    "torch>=2.3.0",
    "torchaudio>=2.3.0",
]
vision = [
    "Pillow>=10.4.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    "qwen-vl-utils>=0.0.8",
]
finetuning = [
    "peft>=0.12.0",
    "datasets>=2.21.0",
    "trl>=0.10.0",
    "bitsandbytes>=0.43.0",
    "unsloth>=2024.9.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "deepeval>=1.5.0",
    "streamlit>=1.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["temporalos"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
